% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_json.R
\name{download_json_data}
\alias{download_json_data}
\title{Fetch Data from EEA API}
\usage{
download_json_data(url, file_path)
}
\arguments{
\item{endpoint}{A character string of the API endpoint}

\item{params}{A list of parameters to pass to the API}
}
\value{
A data.frame or list containing the fetched data
}
\description{
Fetch Data from EEA API
}
\examples{
data <- fetch_eea_data("some_endpoint", list(param1 = "value1", param2 = "value2"))
}
