% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_info_from_json}
\alias{extract_info_from_json}
\title{Fetch Data from EEA API}
\usage{
extract_info_from_json(json_text)
}
\arguments{
\item{endpoint}{A character string of the API endpoint}

\item{params}{A list of parameters to pass to the API}
}
\value{
A data.frame or list containing the fetched data
}
\description{
Fetch Data from EEA API
}
\examples{
data <- fetch_eea_data("some_endpoint", list(param1 = "value1", param2 = "value2"))
}
