% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dataset.R
\name{fetch_data_from_uris}
\alias{fetch_data_from_uris}
\title{Fetch Data from EEA API}
\usage{
fetch_data_from_uris(uris)
}
\arguments{
\item{endpoint}{A character string of the API endpoint}

\item{params}{A list of parameters to pass to the API}
}
\value{
A data.frame or list containing the fetched data
}
\description{
Fetch Data from EEA API
}
\examples{
data <- fetch_eea_data("some_endpoint", list(param1 = "value1", param2 = "value2"))
}
